requirements.txt
nextcord==2.4.2
python-dotenv==1.0.0
PyNaCL==1.4.0
